Extract MarketFaceImpl to separate file
